requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
flet==0.21.2
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
